+++ src/components/files/FileToolbar/FileToolbar.tsx
@@ -1,13 +1,32 @@
+import { useStore } from "effector-react";
 import { FC } from "react";
-import { downloadCurrentFile } from "models/levelsets";
-import { Button } from "ui/button";
+import { $hasOtherFiles, downloadCurrentFile } from "models/levelsets";
+import { Button, ButtonDropdown, Toolbar } from "ui/button";
 import { ColorType } from "ui/types";
-import { svgs } from "ui/icon";
+import { IconStack, IconStackType, svgs } from "ui/icon";
 import { useFileButtonsProps } from "./useFileButtonsProps";
 
-export const FileToolbar: FC = () => {
+const removeOtherStack: IconStack = [[IconStackType.Index, <svgs.Trash />]];
+
+interface Props {
+  isCompact?: boolean;
+}
+
+export const FileToolbar: FC<Props> = ({ isCompact = false }) => {
+  const hasOtherFiles = useStore($hasOtherFiles);
   const { filename, handlers } = useFileButtonsProps();
 
+  const removeButton = (
+    <Button
+      uiColor={ColorType.DANGER}
+      icon={<svgs.Trash />}
+      disabled={!filename}
+      title={filename ? `Remove levelset "${filename}" from memory` : undefined}
+      onClick={handlers?.remove}
+    />
+  );
+
+  const removeOthersTitle = `Remove others but "${filename}"`;
   return (
     <>
       <Button
@@ -30,15 +49,37 @@ export const FileToolbar: FC = () => {
         title="Rename file"
         onClick={handlers?.rename}
       />
-      <Button
-        uiColor={ColorType.DANGER}
-        icon={<svgs.Trash />}
-        disabled={!filename}
-        title={
-          filename ? `Remove levelset "${filename}" from memory` : undefined
-        }
-        onClick={handlers?.remove}
-      />
+
+      {hasOtherFiles && !isCompact ? (
+        <ButtonDropdown
+          standalone={removeButton}
+          buttonProps={{ uiColor: ColorType.DANGER }}
+        >
+          <Toolbar>
+            <Button
+              uiColor={ColorType.DANGER}
+              icon={<svgs.Trash />}
+              iconStack={removeOtherStack}
+              onClick={handlers?.removeOthers}
+            >
+              {removeOthersTitle}
+            </Button>
+          </Toolbar>
+        </ButtonDropdown>
+      ) : (
+        <>
+          {removeButton}
+          {hasOtherFiles && (
+            <Button
+              uiColor={ColorType.DANGER}
+              icon={<svgs.Trash />}
+              iconStack={removeOtherStack}
+              title={removeOthersTitle}
+              onClick={handlers?.removeOthers}
+            />
+          )}
+        </>
+      )}
     </>
   );
 };