+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add: options for SO test page:
   - Plasma limit and Plasma time;
   - Use Serial ports.
+- Add: "Remove Other files" button.
 - Enh: Reduce paddings in several UI components to compress whole UI.
 - Enh: Hide "Append level" button in dropdown as if it's rate used feature.
 - Enh: Hide "Insert/Append/Remove" buttons in SP format, since it's useless.